requirements.txt
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
lxml==4.9.2
numpy==1.24.2
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.10.9
PySocks==1.7.1
pyTelegramBotAPI==4.11.0
pytest==7.3.1
pytest-base-url==2.0.0
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-selenium==4.0.1
pytest-variables==3.0.0
python-dotenv==1.0.0
redis==4.5.4
requests==2.28.2
selenium==4.9.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tenacity==8.2.2
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
urllib3==1.26.14
wsproto==1.2.0